:sparkles:feat:Kakao send message to myself test cleear

+++ src/main/java/excluz/excluz/domain/event/eventApplicant/service/EventApplicantService.java
@@ -102,6 +102,10 @@ public class EventApplicantService {
             throw new BadRequestException(ErrorCode.EVENT_APPLICANT_ALREADY_CONFIRMED);
         }
 
+        if (event.getIsCompleted()) {
+            throw new BadRequestException(ErrorCode.EVENT_ALREADY_CLOSED);
+        }
+
         eventApplicantRepository.delete(eventApplicant);
     }
